Revert "New pkl files in alpha and updated alphamodel.py"
